requirements.txt
flask>=3.0
pystray>=0.19
Pillow>=10.0
